Renaming Moe::gp_parameters, Moe::gp_algorithm
